Refactor Opal OPl3 emulator for easy reusing
