+++ internal/c/parts/audio/build.mk
@@ -13,7 +13,7 @@ MINIAUDIO_REAL_OBJS := $(patsubst %.cpp,$(PATH_INTERNAL_C)/parts/audio/%.o,$(MIN
 MINIAUDIO_STUB_OBJS := $(patsubst %.cpp,$(PATH_INTERNAL_C)/parts/audio/%.o,$(MINIAUDIO_STUB_SRCS))
 
 ifdef DEP_AUDIO_MINIAUDIO
-	MINIAUDIO_OBJS :=  $(MINIAUDIO_REAL_OBJS) $(MA_VTABLES_OBJS) $(HIVELY_OBJS) $(LIBXMP_LIB)
+	MINIAUDIO_OBJS :=  $(MINIAUDIO_REAL_OBJS) $(MA_VTABLES_OBJS) $(HIVELY_OBJS) $(OPAL_OBJS) $(LIBXMP_LIB)
 
 ifdef DEP_AUDIO_DECODE_MIDI
 	MINIAUDIO_OBJS += $(MIDI_MA_VTABLES_OBJS)
